requirements.txt
lmstudio==1.3.0
PyAudio==0.2.14
python-dotenv==1.1.0
pyttsx3==2.98
Requests==2.32.3
SpeechRecognition==3.10.0
